~/.bashrc - correction for copyconfigs function. only specify the filename itself as the copy destination if we're copying a file. otherwise the copied directory ends up inside the directory at the new location instead of overwriting it

+++ .bashrc
@@ -93,7 +93,11 @@ copyconfigs() {
         echo "Changed: $HOME/$x"
       else
         echo "Copying: $HOME/$x"
-        cp -r "$HOME/$x" "$HOME/Devel/GitHub/my-config/$x"
+        if [ -f "$x" ]; then 
+          cp -r "$HOME/$x" "$HOME/Devel/GitHub/my-config/$x"
+        else
+          cp -r "$HOME/$x" "$HOME/Devel/GitHub/my-config"
+        fi
       fi
     fi
   done
